+++ backend/app/models/script.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, ForeignKey, JSON
+from sqlalchemy import Column, Integer, BigInteger, String, Text, ForeignKey, JSON
 from sqlalchemy.orm import relationship
 from app.models.base import Base, TimestampMixin
 
@@ -6,8 +6,8 @@ from app.models.base import Base, TimestampMixin
 class Script(Base, TimestampMixin):
     __tablename__ = "scripts"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
+    user_id = Column(BigInteger, ForeignKey("users.id"), nullable=False, index=True)
     title = Column(String(200), nullable=True)
     content = Column(Text, nullable=False, comment="脚本内容")
     output_type = Column(String(20), nullable=False, comment="输出类型: image_set/single_image/video")
