+++ backend/app/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, Numeric, Text
+from sqlalchemy import Column, Integer, BigInteger, String, Boolean, DateTime, Numeric, Text
 from sqlalchemy.orm import relationship
 from app.models.base import Base, TimestampMixin
 
@@ -6,7 +6,7 @@ from app.models.base import Base, TimestampMixin
 class User(Base, TimestampMixin):
     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True, index=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, index=True, autoincrement=True)
     email = Column(String(100), unique=True, nullable=True, index=True)
     phone = Column(String(20), unique=True, nullable=True, index=True)
     password = Column(String(255), nullable=False)
@@ -15,7 +15,7 @@ class User(Base, TimestampMixin):
     status = Column(Integer, default=1, nullable=False, comment="状态: 0禁用 1正常")
     balance = Column(Numeric(10, 2), default=0.00, nullable=False)
     invite_code = Column(String(20), unique=True, nullable=True, index=True)
-    invited_by = Column(Integer, nullable=True, index=True)
+    invited_by = Column(BigInteger, nullable=True, index=True)
     last_login_at = Column(DateTime, nullable=True)
     last_login_ip = Column(String(50), nullable=True)
